requirements.txt
# AI / LLM
anthropic>=0.39.0 # Keep for future use (Prod env)
google-genai>=1.63.0 # Gemini API client library

# Data Processing
pandas>=2.2.0

# CLI Interface
click>=8.1.7
rich>=13.7.0

# Environment Variables
python-dotenv>=1.0.0